--- src/main/java/io/github/hildi/can/service/StandardJavaSerializationService.java
@@ -1,76 +1,0 @@
-package io.github.hildi.can.service;
-
-import io.github.hildi.can.exceptions.*;
-import io.github.hildi.can.model.User;
-
-import java.io.*;
-
-/**
- * Created by Serhii Hildi on 15.03.19.
- */
-public class StandardJavaSerializationService implements SerializationService {
-
-    @Override
-    public void serialize(User user, File file) {
-
-        assertParamNotNull(user, file);
-        assertFileExist(file);
-        assertFileWritable(file);
-
-        try (ObjectOutputStream out = new ObjectOutputStream(new FileOutputStream(file))) {
-            out.writeObject(user);
-        } catch (FileNotFoundException e) {
-            throw new SerializationException("Failed to serialize data, reason: file " + file.getName() + " is not found. ", e);
-        } catch (IOException e) {
-            throw new SerializationException("Failed to serialize data. " + file.getName(), e);
-        }
-    }
-
-    @Override
-    public User deserialize(File file) {
-
-        assertFileNotNull(file);
-        assertFileReadable(file);
-
-        try (ObjectInputStream in = new ObjectInputStream(new FileInputStream(file))) {
-            return (User) in.readObject();
-        } catch (FileNotFoundException e) {
-            throw new SerializationException("Failed to deserialize data, reason: file " + file.getName() + " is not found. ", e);
-        } catch (ClassNotFoundException e) {
-            throw new SerializationException("Failed to deserialize data, reason: class is not found. " + e.getMessage(), e);
-        } catch (IOException e) {
-            throw new SerializationException("Failed to deserialize data. " + e.getMessage(), e);
-        }
-    }
-
-    static void assertFileExist(File file) {
-        if (!file.exists()) {
-            throw new IllegalArgumentException("Failed to serialize, reason: file " + file.getName() + " doesn't exist.");
-        }
-    }
-
-    static void assertFileWritable(File file) {
-        if (!file.canWrite()) {
-            throw new NotWritableFileException("Failed to serialize, reason: " + file.getName() + " is not writable.");
-        }
-    }
-
-    static void assertFileReadable(File file) {
-        if (!file.canRead()) {
-            throw new NotReadableFileException("Failed to deserialize, reason: " + file.getName() + " is not readable or doesn't exist.");
-        }
-    }
-
-    static void assertParamNotNull(User user, File file) {
-        if (user == null) {
-            throw new IllegalArgumentException("Failed to serialize, reason: User is null.");
-        }
-        assertFileNotNull(file);
-    }
-
-    static void assertFileNotNull(File file) {
-        if (file == null) {
-            throw new IllegalArgumentException("Failed to serialize, reason: File is null.");
-        }
-    }
-}
